==> Metabolomics/scripts/01_LoadData_COF_visitAB_metabolomics.Rmd <==
---
title: "Metabolomics Overfeeding Analysis"
subtitle: "Load and Save Raw Data"
author: "Sofie Olund Villumsen"
date: "2023"
output: html_document
---

# Initial settings
Clear workspace
```{r}
rm(list = ls()) 
```

Load packages
```{r message=FALSE}
library(tidyverse)
library(readxl)
```

Set wd
```{r setup}
path <- "L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/github/COFpaper_omics_adiposetissue/Metabolomics"

knitr::opts_knit$set(root.dir = path)
```

Load data
```{r message=FALSE}
metabo_data <- read_xlsx(paste(path, "data/raw_data/Kopi_af_Overfeed_Metablomics_Final_Output_Simple.xlsx", sep="/")) %>% 
  write_csv(paste(path, "data/01_metabolomics_data.csv", sep="/"))

clinical_data <- read_xlsx(paste(path, "data/raw_data/hcot_clinical_data.xlsx", sep="/")) %>% 
  write_csv(paste(path, "data/01_clinical_data.csv", sep="/"))
```
